Changed the way the images are painted and started improving GUI code.
Dont think it makes much sense the way the images were painted earlier, since when you had a map that was for example width = 20 , height = 5, instead of squares the tiles would be rectangles, which is really ugly.
This way its much easier to do calculations of position of mouse and the look of it is better.

Made MapCreator window not resizable because of this.